Use linker sets for creating a device table.
Use a linker set to create the device table. See devtabreset for example usage.
This is how the cool kids like to do it now. But is it too gross?

The alignment of the struct and in the kernel.ld is required or gcc won't get it right.
I've tried.

I'm actually ok with NOT doing this if people think it's gross. We can just continue
to use the PRE rules in core.json and keep the awk scripts -- I'm fine with that too.
Just let me know, and if you don't like this, I can strip this commit.

BTW, this type of thing is one reason we might want to get gerritub.io going.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>